[Spring 지하철 노선도 - 3단계] 더즈(이동규) 미션 제출합니다. (#262)
* refactor: ControllerAdvice로 이름 변경 및 예외 메시지 추가

* test: Station, Line 테스트 코드 추가

* feat: SECTION 테이블 추가

* test: 인수 테스트 중복 코드 제거

* test: RestAssuredUtil 추가 후 중복 로직 제거

* feat: 노선을 등록할 때 구간도 함께 등록하는 기능 구현

* refactor: LineService가 LineDao만 의존하도록 변경

* refactor: 안쓰는 메서드 삭제

* refactor: 사용하지 않는 생성자 제거

* feat: Sections에서 정렬된 지하철역 반환 기능 구현

* test: LineDaoTest 노선 조회 테스트 추가

* feat: 구간 추가 기능 구현

* test: 지하철 구간 추가 인수 테스트 추가

* refactor: 구간 추가 코드 리팩터링

* test: Section.isIncludedIn 테스트 추가

* refactor: 안쓰는 메서드 삭제

* feat: 지하철 구간 삭제 기능 구현

* chore: 사용하지 않는 import 삭제

* refactor: 역 정렬 구현 로직 변경

* refactor: 구간 삭제 메서드 분리

* refactor: 정렬 로직 최적화

* fix: Line, Section에 외래키 제약 추가

* refactor: 인수 테스트 격리를 위한 sql, delete -> truncate 사용

* style: 끝줄 개행 없는 파일 개행 추가

* test: StationAcceptanceTest 검증 구체화 및 테스트 케이스 추가

* test: LineAcceptanceTest 검증 구체화 및 테스트 케이스 추가

* test: SectionAcceptanceTest 검증 구체화 및 테스트 케이스 추가

* test: StationService의 findOne() 테스트 케이스 추가

* style: listStations -> findAllStations 메서드 이름 변경

* test: LineServiceTest 검증 구체화

* refactor: 구간 추가 메서드 이름 및 구조 변경

* refactor: SectionRepository 구현

* refactor: LineRepository 구현

* fix: 구간으로 등록된 역을 삭제할 수 없도록 변경

* refactor: SectionsDirtyChecker 적용 후 Sections의 add, delete 메서드 void로 변경

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import wooteco.subway.dao.StationDao;
+import wooteco.subway.dao.repository.SectionRepository;
 import wooteco.subway.domain.Station;
 
 @Service
@@ -13,9 +14,11 @@ import wooteco.subway.domain.Station;
 public class StationService {
 
 	private final StationDao stationDao;
+	private final SectionRepository sectionRepository;
 
-	public StationService(StationDao stationDao) {
+	public StationService(StationDao stationDao, SectionRepository sectionRepository) {
 		this.stationDao = stationDao;
+		this.sectionRepository = sectionRepository;
 	}
 
 	@Transactional
@@ -31,12 +34,19 @@ public class StationService {
 		}
 	}
 
-	public List<Station> listStations() {
+	public List<Station> findAllStations() {
 		return stationDao.findAll();
 	}
 
 	@Transactional
 	public void remove(Long id) {
+		if (sectionRepository.existByStation(id)) {
+			throw new IllegalArgumentException("구간으로 등록되어 있어 삭제할 수 없습니다.");
+		}
 		stationDao.remove(id);
+	}
+
+	public Station findOne(Long id) {
+		return stationDao.findById(id);
 	}
 }
